refactor: split FFN class into many subclasses
